// src/generated/res/gtfs-realtime.ts
export interface FeedMessage {
  /** Metadata about this feed and feed message. */
  header:
    | FeedHeader
    | undefined;
  /** Contents of the feed. */
  entity: FeedEntity[];
}

/** Metadata about a feed, included in feed messages. */
export interface FeedHeader {
  /**
   * Version of the feed specification.
   * The current version is 2.0.  Valid versions are "2.0", "1.0".
   */
  gtfsRealtimeVersion: string;
  incrementality?:
    | FeedHeader_Incrementality
    | undefined;
  /**
   * This timestamp identifies the moment when the content of this feed has been
   * created (in server time). In POSIX time (i.e., number of seconds since
   * January 1st 1970 00:00:00 UTC).
   */
  timestamp?:
    | number
    | undefined;
  /**
   * String that matches the feed_info.feed_version from the GTFS feed that the real
   * time data is based on. Consumers can use this to identify which GTFS feed is
   * currently active or when a new one is available to download.
   */
  feedVersion?: string | undefined;
}

/**
 * Determines whether the current fetch is incremental.  Currently,
 * DIFFERENTIAL mode is unsupported and behavior is unspecified for feeds
 * that use this mode.  There are discussions on the GTFS Realtime mailing
 * list around fully specifying the behavior of DIFFERENTIAL mode and the
 * documentation will be updated when those discussions are finalized.
 */
export enum FeedHeader_Incrementality {
  FULL_DATASET = 0,
  DIFFERENTIAL = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface FeedEntity {
  /**
   * The ids are used only to provide incrementality support. The id should be
   * unique within a FeedMessage. Consequent FeedMessages may contain
   * FeedEntities with the same id. In case of a DIFFERENTIAL update the new
   * FeedEntity with some id will replace the old FeedEntity with the same id
   * (or delete it - see is_deleted below).
   * The actual GTFS entities (e.g. stations, routes, trips) referenced by the
   * feed must be specified by explicit selectors (see EntitySelector below for
   * more info).
   */
  id: string;
  /**
   * Whether this entity is to be deleted. Relevant only for incremental
   * fetches.
   */
  isDeleted?:
    | boolean
    | undefined;
  /**
   * Data about the entity itself. Exactly one of the following fields must be
   * present (unless the entity is being deleted).
   */
  tripUpdate?: TripUpdate | undefined;
  vehicle?: VehiclePosition | undefined;
  alert?:
    | Alert
    | undefined;
  /** NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future. */
  shape?: Shape | undefined;
  stop?: Stop | undefined;
  tripModifications?: TripModifications | undefined;
}
// ...
export interface TripUpdate {
  /**
   * The Trip that this message applies to. There can be at most one
   * TripUpdate entity for each actual trip instance.
   * If there is none, that means there is no prediction information available.
   * It does *not* mean that the trip is progressing according to schedule.
   */
  trip:
    | TripDescriptor
    | undefined;
  /** Additional information on the vehicle that is serving this trip. */
  vehicle?:
    | VehicleDescriptor
    | undefined;
  /**
   * Updates to StopTimes for the trip (both future, i.e., predictions, and in
   * some cases, past ones, i.e., those that already happened).
   * The updates must be sorted by stop_sequence, and apply for all the
   * following stops of the trip up to the next specified one.
   *
   * Example 1:
   * For a trip with 20 stops, a StopTimeUpdate with arrival delay and departure
   * delay of 0 for stop_sequence of the current stop means that the trip is
   * exactly on time.
   *
   * Example 2:
   * For the same trip instance, 3 StopTimeUpdates are provided:
   * - delay of 5 min for stop_sequence 3
   * - delay of 1 min for stop_sequence 8
   * - delay of unspecified duration for stop_sequence 10
   * This will be interpreted as:
   * - stop_sequences 3,4,5,6,7 have delay of 5 min.
   * - stop_sequences 8,9 have delay of 1 min.
   * - stop_sequences 10,... have unknown delay.
   */
  stopTimeUpdate: TripUpdate_StopTimeUpdate[];
  /**
   * The most recent moment at which the vehicle's real-time progress was measured
   * to estimate StopTimes in the future. When StopTimes in the past are provided,
   * arrival/departure times may be earlier than this value. In POSIX
   * time (i.e., the number of seconds since January 1st 1970 00:00:00 UTC).
   */
  timestamp?:
    | number
    | undefined;
  /**
   * The current schedule deviation for the trip.  Delay should only be
   * specified when the prediction is given relative to some existing schedule
   * in GTFS.
   *
   * Delay (in seconds) can be positive (meaning that the vehicle is late) or
   * negative (meaning that the vehicle is ahead of schedule). Delay of 0
   * means that the vehicle is exactly on time.
   *
   * Delay information in StopTimeUpdates take precedent of trip-level delay
   * information, such that trip-level delay is only propagated until the next
   * stop along the trip with a StopTimeUpdate delay value specified.
   *
   * Feed providers are strongly encouraged to provide a TripUpdate.timestamp
   * value indicating when the delay value was last updated, in order to
   * evaluate the freshness of the data.
   *
   * NOTE: This field is still experimental, and subject to change. It may be
   * formally adopted in the future.
   */
  delay?: number | undefined;
  tripProperties?: TripUpdate_TripProperties | undefined;
}
// ...
export const FeedMessage: MessageFns<FeedMessage> = {
  encode(message: FeedMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      FeedHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    for (const v of message.entity) {
      FeedEntity.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FeedMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFeedMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = FeedHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.entity.push(FeedEntity.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FeedMessage {
    return {
      header: isSet(object.header) ? FeedHeader.fromJSON(object.header) : undefined,
      entity: globalThis.Array.isArray(object?.entity) ? object.entity.map((e: any) => FeedEntity.fromJSON(e)) : [],
    };
  },

  toJSON(message: FeedMessage): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = FeedHeader.toJSON(message.header);
    }
    if (message.entity?.length) {
      obj.entity = message.entity.map((e) => FeedEntity.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FeedMessage>, I>>(base?: I): FeedMessage {
    return FeedMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FeedMessage>, I>>(object: I): FeedMessage {
    const message = createBaseFeedMessage();
    message.header = (object.header !== undefined && object.header !== null)
      ? FeedHeader.fromPartial(object.header)
      : undefined;
    message.entity = object.entity?.map((e) => FeedEntity.fromPartial(e)) || [];
    return message;
  },
};

function createBaseFeedHeader(): FeedHeader {
  return { gtfsRealtimeVersion: "", incrementality: 0, timestamp: 0, feedVersion: "" };
}

export const FeedHeader: MessageFns<FeedHeader> = {
  encode(message: FeedHeader, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.gtfsRealtimeVersion !== "") {
      writer.uint32(10).string(message.gtfsRealtimeVersion);
    }
    if (message.incrementality !== undefined && message.incrementality !== 0) {
      writer.uint32(16).int32(message.incrementality);
    }
    if (message.timestamp !== undefined && message.timestamp !== 0) {
      writer.uint32(24).uint64(message.timestamp);
    }
    if (message.feedVersion !== undefined && message.feedVersion !== "") {
      writer.uint32(34).string(message.feedVersion);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FeedHeader {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFeedHeader();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.gtfsRealtimeVersion = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.incrementality = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.timestamp = longToNumber(reader.uint64());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.feedVersion = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FeedHeader {
    return {
      gtfsRealtimeVersion: isSet(object.gtfsRealtimeVersion)
        ? globalThis.String(object.gtfsRealtimeVersion)
        : isSet(object.gtfs_realtime_version)
        ? globalThis.String(object.gtfs_realtime_version)
        : "",
      incrementality: isSet(object.incrementality) ? feedHeader_IncrementalityFromJSON(object.incrementality) : 0,
      timestamp: isSet(object.timestamp) ? globalThis.Number(object.timestamp) : 0,
      feedVersion: isSet(object.feedVersion)
        ? globalThis.String(object.feedVersion)
        : isSet(object.feed_version)
        ? globalThis.String(object.feed_version)
        : "",
    };
  },

  toJSON(message: FeedHeader): unknown {
    const obj: any = {};
    if (message.gtfsRealtimeVersion !== "") {
      obj.gtfsRealtimeVersion = message.gtfsRealtimeVersion;
    }
    if (message.incrementality !== undefined && message.incrementality !== 0) {
      obj.incrementality = feedHeader_IncrementalityToJSON(message.incrementality);
    }
    if (message.timestamp !== undefined && message.timestamp !== 0) {
      obj.timestamp = Math.round(message.timestamp);
    }
    if (message.feedVersion !== undefined && message.feedVersion !== "") {
      obj.feedVersion = message.feedVersion;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FeedHeader>, I>>(base?: I): FeedHeader {
    return FeedHeader.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FeedHeader>, I>>(object: I): FeedHeader {
    const message = createBaseFeedHeader();
    message.gtfsRealtimeVersion = object.gtfsRealtimeVersion ?? "";
    message.incrementality = object.incrementality ?? 0;
    message.timestamp = object.timestamp ?? 0;
    message.feedVersion = object.feedVersion ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}